4_Model_Deployment_Using_Cloud_Services/ml_ops_v1/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
Flask
werkzeug
joblib
scikit-learn
numpy
